Fixed a bug when len(snake_list) = 2
When there are only two segments in the snake_list, including the head, since the segments are always following the segment before, the head will never come in contact with the second segment because it's always trailing. To fix that we changed how the movement works. Instead of having all four movements available at all times, we changed it so that you cannot double back over yourself, which would cause a game over state.